+++ mobile/android/android-components/components/lib/crash/src/test/java/mozilla/components/lib/crash/service/SendCrashReportServiceTest.kt
@@ -98,6 +98,7 @@ class SendCrashReportServiceTest {
                 ),
                 scope = scope,
                 notificationsDelegate = mock(),
+                crashBuilder = mock(),
             ),
         ).install(testContext)
         val originalCrash = Crash.NativeCodeCrash(