+++ mobile/android/fenix/app/src/main/java/org/mozilla/fenix/components/Analytics.kt
@@ -9,6 +9,7 @@ import android.app.PendingIntent
 import android.content.Context
 import android.content.Intent
 import android.os.Build
+import mozilla.components.lib.crash.CrashBuilder
 import mozilla.components.lib.crash.CrashReporter
 import mozilla.components.lib.crash.sentry.SentryService
 import mozilla.components.lib.crash.service.CrashReporterService
@@ -122,6 +123,7 @@ class Analytics(
             enabled = true,
             nonFatalCrashIntent = pendingIntent,
             notificationsDelegate = context.components.notificationsDelegate,
+            crashBuilder = CrashBuilder(runtimeTagProviders = emptyList())
         )
     }