+++ mobile/android/fenix/app/src/test/java/org/mozilla/fenix/components/metrics/BreadcrumbRecorderTest.kt
@@ -58,6 +58,7 @@ class BreadcrumbRecorderTest {
                 services = listOf(service),
                 shouldPrompt = CrashReporter.Prompt.NEVER,
                 notificationsDelegate = mockk(),
+                crashBuilder = mockk(),
             ),
         )